Updates for #1193     - The nightly build is specifying a higher level of optimization through to the compiler than the debug mode that I have been using for my testing. Building calwf3 with the lower optimization produced no errors or warnings and ran through cleanly, but the high optimization brought on a segfault inside the CTE code in the first openmp section. This was only happening on the Linux cluster, the MAC builds showed no issue. I went back 2 versions of the wf3cte.c code and redid the changes by hand step by step. The problem area seemed to be a set of arrays which would rather be doubles than floats. I also changed the remaining floats to doubles, where I could, and removed more of the memcpy statements, making them regular array assignments. These things seem to have cleared the issues, and hopefully preventing new ones for now.
    - I also removed a superfluous openmp print statement from maincte.c and cleaned up some more informational print statements.

    - I added a time measurement, with verbose on the code prints how long the CTE section took to run, with the specification of number of threads/cpus.
    



git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/hstcal/trunk@45807 fe389314-cf27-0410-b35b-8c050e845b92

+++ pkg/wfc3/calwf3/wf3cte/cte.h
@@ -10,19 +10,19 @@ typedef struct {
     int cte_len; /*max length of cte trail */
     double   rn_amp; /*read noise amplitude for clipping */
     int n_forward; /* number of forward modeling iterations */
-    int n_par; /*numver of iterations in parallel transfer */
-    float scale_frac; /*scaling of cte model relative to ctedate1*/
+    int n_par; /*number of iterations in parallel transfer */
+    double scale_frac; /*scaling of cte model relative to ctedate1*/
     int noise_mit; /*read noise mitigation algorithm*/
-    float thresh; /*over subtraction threshold*/        
+    double thresh; /*over subtraction threshold*/        
     int cte_traps; /*number of valid TRAPS in file for reallocation*/
     int wcol_data[TRAPS]; /*trap number, insync with number of traps*/
-    float qlevq_data[TRAPS];/*charge packet size in electrons*/
-    float dpdew_data[TRAPS];/*trap size in electrons*/  
+    double qlevq_data[TRAPS];/*charge packet size in electrons*/
+    double dpdew_data[TRAPS];/*trap size in electrons*/  
     int   iz_data[RAZ_COLS]; /*column number in raz format*/
-    float scale512[RAZ_COLS]; /*scaling appropriate at row 512 */
-    float scale1024[RAZ_COLS];/*scaling appropriate at row 1024 */
-    float scale1536[RAZ_COLS];/*scaling appropriate at row 1536 */
-    float scale2048[RAZ_COLS];/*scaling appropriate at row 2048 */
+    double scale512[RAZ_COLS]; /*scaling appropriate at row 512 */
+    double scale1024[RAZ_COLS];/*scaling appropriate at row 1024 */
+    double scale1536[RAZ_COLS];/*scaling appropriate at row 1536 */
+    double scale2048[RAZ_COLS];/*scaling appropriate at row 2048 */
     char descrip2[SZ_LINE]; /*descrip from table row, not read in for cte purposes*/
     int fix_rocr; /*make allowance for readout cosmic rays*/
     FloatHdrData *rprof; /*differential trail profile as image*/
@@ -78,7 +78,7 @@ int findPreScanBias(SingleGroup *, float *, float *);
 int find_dadj(int ,int , double [][RAZ_ROWS], double [][RAZ_ROWS], double , double *);
 int rsz2rsc(WF3Info *, SingleGroup *, SingleGroup *, CTEParams * );
 int inverse_cte_blur(SingleGroup *, SingleGroup *, SingleGroup *, CTEParams *, int, double);
-int sim_colreadout_l(float *, float *, float *, CTEParams *);
+int sim_colreadout_l(double *, double *, double *, CTEParams *);
 int CompareCTEParams(SingleGroup *, CTEParams *);
 int cteHistory (WF3Info *, Hdr *);
 int free_array(float **ptr, int rows, int columns);
